[Feat] Add a popup window function.
